+++ lib/kitchen/driver/ssh_base.rb
@@ -1,13 +1,8 @@
-#
-# Author:: Fletcher Nichol (<fnichol@nichol.ca>)
-#
-# Copyright (C) 2012, Fletcher Nichol
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+#    https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -303,7 +298,7 @@ module Kitchen
         pseudo_state[:username] = username if username
         pseudo_state.merge!(options)
 
-        instance.transport.connection(backcompat_merged_state(pseudo_state))
+        instance.transport.connection(**backcompat_merged_state(pseudo_state))
           .wait_until_ready
       end