+++ lib/kitchen/ssh.rb
@@ -7,7 +7,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+#    https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -92,7 +92,7 @@ module Kitchen
     # @param remote [String] path to remote file destination
     # @param options [Hash] configuration options that are passed to
     #   `Net::SCP.upload`
-    # @see http://net-ssh.github.io/net-scp/classes/Net/SCP.html#method-i-upload
+    # @see https://net-ssh.github.io/net-scp/classes/Net/SCP.html#method-i-upload
     def upload!(local, remote, options = {}, &progress)
       require "net/scp" unless defined?(Net::SCP)
       if progress.nil?
@@ -124,7 +124,7 @@ module Kitchen
     # @param options [Hash] configuration options that are passed to
     #   `Net::SCP.upload`
     # @option options [true,false] :recursive recursive copy (default: `true`)
-    # @see http://net-ssh.github.io/net-scp/classes/Net/SCP.html#method-i-upload
+    # @see https://net-ssh.github.io/net-scp/classes/Net/SCP.html#method-i-upload
     def upload_path!(local, remote, options = {}, &progress)
       options = { recursive: true }.merge(options)