+++ spec/kitchen/data_munger_spec.rb
@@ -7,7 +7,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+#    https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -2607,7 +2607,7 @@ module Kitchen # rubocop:disable Metrics/ModuleLength
           )
         end
 
-        it "merges old common busser name to version with exising verifier" do
+        it "merges old common busser name to version with existing verifier" do
           result = DataMunger.new(
             {
               busser: "starship",