+++ spec/kitchen/lazy_hash_spec.rb
@@ -7,7 +7,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+#    https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -26,7 +26,7 @@ describe Kitchen::LazyHash do
 
   let(:hash_obj) do
     {
-      shed_color: ->(c) { c.color }, # rubocop:disable Style/SymbolProc
+      shed_color: lambda(&:color),
       barn: "locked",
       genre: proc { |c| "#{c.metal} metal" },
     }
@@ -105,8 +105,7 @@ describe Kitchen::LazyHash do
     end
 
     it "yields each item to the block if a block is given to each()" do
-      items = []
-      Kitchen::LazyHash.new(hash_obj, context).each { |i| items << i } # rubocop:disable Style/MapIntoArray
+      items = Kitchen::LazyHash.new(hash_obj, context).map { |i| i }
       _(items).must_equal [[:shed_color, "blue"], [:barn, "locked"], [:genre, "heavy metal"]]
     end
   end