+++ spec/kitchen/platform_filter_spec.rb
@@ -7,7 +7,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+#    https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -22,44 +22,44 @@ require "kitchen/platform_filter"
 describe Kitchen::PlatformFilter do
   describe ".convert" do
     it "returns an array if a scalar is passed" do
-      Kitchen::PlatformFilter.convert(nil).must_equal []
-      Kitchen::PlatformFilter.convert("string").must_equal [Kitchen::PlatformFilter.new("string")]
+      _(Kitchen::PlatformFilter.convert(nil)).must_equal []
+      _(Kitchen::PlatformFilter.convert("string")).must_equal [Kitchen::PlatformFilter.new("string")]
     end
 
     it "just wraps simple strings" do
-      Kitchen::PlatformFilter.convert(%w{CentOS windows-2016}).must_equal [Kitchen::PlatformFilter.new("CentOS"),
+      _(Kitchen::PlatformFilter.convert(%w{CentOS windows-2016})).must_equal [Kitchen::PlatformFilter.new("CentOS"),
                                                                            Kitchen::PlatformFilter.new("windows-2016")]
     end
 
-    it "converts regexp-like strings into Regexp before wraping" do
-      Kitchen::PlatformFilter.convert(%w{/^win/ /win$/}).must_equal [Kitchen::PlatformFilter.new(/^win/),
+    it "converts regexp-like strings into Regexp before wrapping" do
+      _(Kitchen::PlatformFilter.convert(%w{/^win/ /win$/})).must_equal [Kitchen::PlatformFilter.new(/^win/),
                                                                      Kitchen::PlatformFilter.new(/win$/)]
     end
 
     it "supports IgnoreCase Regexp option" do
-      Kitchen::PlatformFilter.convert("/win/i").must_equal [Kitchen::PlatformFilter.new(/win/i)]
+      _(Kitchen::PlatformFilter.convert("/win/i")).must_equal [Kitchen::PlatformFilter.new(/win/i)]
     end
 
     it "supports Extended Regexp option" do
-      Kitchen::PlatformFilter.convert("/win/x").must_equal [Kitchen::PlatformFilter.new(/win/x)]
+      _(Kitchen::PlatformFilter.convert("/win/x")).must_equal [Kitchen::PlatformFilter.new(/win/x)]
     end
 
     it "supports combination of IgnoreCase & Extended Regexp options" do
-      Kitchen::PlatformFilter.convert(%w{/win/ix /win/xi}).must_equal [Kitchen::PlatformFilter.new(/win/ix),
+      _(Kitchen::PlatformFilter.convert(%w{/win/ix /win/xi})).must_equal [Kitchen::PlatformFilter.new(/win/ix),
                                                                        Kitchen::PlatformFilter.new(/win/ix)]
     end
   end
 
   describe ".new" do
     it "raises an ArgumentError if value is neither a string nor Regexp" do
-      proc { Kitchen::PlatformFilter.new(Object.new) }.must_raise ::ArgumentError
+      _ { Kitchen::PlatformFilter.new(Object.new) }.must_raise ::ArgumentError
     end
   end
 
   describe "#value" do
     it "returns the original value" do
-      Kitchen::PlatformFilter.new("string").value.must_equal "string"
-      Kitchen::PlatformFilter.new(/regexp/).value.must_equal(/regexp/)
+      _(Kitchen::PlatformFilter.new("string").value).must_equal "string"
+      _(Kitchen::PlatformFilter.new(/regexp/).value).must_equal(/regexp/)
     end
   end
 end