+++ spec/kitchen/provisioner/shell_spec.rb
@@ -7,7 +7,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#    http://www.apache.org/licenses/LICENSE-2.0
+#    https://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -444,7 +444,7 @@ describe Kitchen::Provisioner::Shell do
       it "invokes the bootstrap.ps1 script" do
         config[:root_path] = '\\r'
 
-        _(cmd).must_match regexify(%{& \\r\\bootstrap.ps1})
+        _(cmd).must_match regexify(%{& "\\r\\bootstrap.ps1"})
       end
     end
   end